Some basic types.
Wares aren't going anywhere, of course.

WarehouseAddr (also previously referred to as WarehouseCoords) is a placeholder at this moment.

WarehouseDialer is a factory pattern that... well, might be too much enterprise java, but we'll see.  I'm trying to sketch out all the major pieces at once...

Justification is a placeholder at the moment.  It's meant to represent some records about why a ware is worth storing: we'll need some structures like this for reasonable garbage collection to be implementable.

WarehouseAgent is the meaty bit.  The core realization there is it only ever exposes a Put method.  *It doesn't have a Get method* -- the symmetry of that always seems obvious, and tempting, but is a trap.

Not yet shown: some interfaces for putting things on the filesystem, unpacked.  Those, too, will have a Put method... and it will take a list of WarehouseAgent or WarehouseAddr.  This is the secret to why WarehouseAgent itself never has a Get method.

Signed-off-by: Eric Myhre <hash@exultant.us>