Propose more args to transmat methods to describe how concurrency is going to go.
We're *not* returning promises; that always leaves you with a bunch of questionmarks and additional layers of helper wrapper methods.  Instead, we're passing down context for cancellability, and optional "monitor" objects, which will get callbacks from the same goroutine that started the call.  Concurrency is left in the hands of the caller; the whole thing behaves blockingly.

Signed-off-by: Eric Myhre <hash@exultant.us>