Begin defining standard file metadata.
This is... almost the exact same struct as the tar headers.  Using the tar headers in the repeatr source to date has served pretty well.  Our changes are mostly a matter of stripping things, in fact.  But having it properly standardized in a type of our own maintaining seems wise.

Even if it does cause us to have some irritating memshuffle dances for the tar transmat in particular.

Signed-off-by: Eric Myhre <hash@exultant.us>