Port over the fshash bucket stuff.
Treewalk lib comes too, no patching required.

There's a few things that feel a little crotchety in this fshash bucket implementation, but overall the codesmells from this have been fairly low over time, so we'll keep it; some smoothing later is possible, but shouldn't be a collosal breakage thing.

The filename strings are getting duplicated now as of the port, to compensate for the fact we're mostly operating in terms of RelPath objects, and our hashing and sorting and traversal standards believe pretty fiercely in the trailing-slash-for-dirs convention.  This seems not ideal for performance, but also... yeah, we'll survive.

Signed-off-by: Eric Myhre <hash@exultant.us>