Chtimes methods! for linux.  Placefile and the fsOp package now compile!
Other platforms pending.  It's possible to support lchtimes completely on some platforms; not on others.  However, we should include the function definitions even on platforms that cannot support the operations: it's better to be able to compile, even if all we can do is a hardcoded error, given that many other features of rio may still work correctly.

Signed-off-by: Eric Myhre <hash@exultant.us>