Add skipChown option to PlaceFile.
Not 100% confident this is the most sensible API here; tossing a booling on to a method sig always seems like a hack.  However, it does seem to do what we mean: it alters whether or not the system needs CAP_CHOWN, and also whether or not it DTRT when the user explicitly wants to unpack as their current UID/GID.

Merely altering the behavior of the function based on whether or not the capability is available would be much less correct: both because it would be comically inefficient to recheck every time, and also because, while that may be a userland/CLI-level indication of "what user wants", it's *very* ugly to let that variable slip *silently* through side channels to this low of a level.

Signed-off-by: Eric Myhre <hash@exultant.us>