A goodly bit of the file transport, which will shortly be usable by the tar transmat.
It seems likely that there will be SOME reusable interfaces for warehouse control and transport, but it's not really clear where exactly those will break down, so the package structure here is leaving room for it in a hopeful way.

(Probably: many things will be key-value blobstorage warehouses, and all those will share some interfaces; and everything else is a freeforall of their own personal complexity.  But we'll see.)

Compared to the previous generation of repeatr transmat code: this major separation between warehouse control and transports vs the packing system is going to result in very different code.  So far it's already looking much cleaner.

Signed-off-by: Eric Myhre <hash@exultant.us>