Belay that last.  Standardize timezones at the edges.
I still think the os.Stat behavior of returning a local timezone is *odd*, to say the least.  But it's more correct to keep those semantics at the `fs` interface layer, and make sure we normalize to UTC as necessary at the transmat layers, I think, just in case we ever encounter a parallel universe where kernels and filesystems *do* sometimes understand and retain TZ data.

Signed-off-by: Eric Myhre <hash@exultant.us>