Assert UID and GID.
This is a funny one: my tests on localhost were passing, but raceway has reds for the last few commits.  Why?  Because the chtimes call is aimed at a file that has a different owner.  My host system uid is still 7000; in the raceway repeatr container, it's 1000. Yep.

Be very careful, always, with what you thing is test coverage when touching the OS...

Signed-off-by: Eric Myhre <hash@exultant.us>