Assert on basedir properties, even on tar fixtures that lack them.
And fix that to actually work again.  (We skipped a lot of the detail work earlier in the porting process, apparently.  More than I'd realized.  Sheesh.  I'm kind of sad that so much of it is resurfacing; I was hoping the better bottom layer of abstractions this time around was going to vanish a larger amount of this fluffery.)

Signed-off-by: Eric Myhre <hash@exultant.us>