Add FS.ReadDirNames; Port over fs.Walk behavior.
Our definition of readdirnames is a little different than standardlib, because we're making it a one-step operation here... meaning it *always* has the n=-1 behavior, since there's no handle for iterability.  We could *also* add one on the `File` interface, but... we'll do that if we, well, care someday.

We should probably change it to do sorted order.  Or at least, the walker probably should.  You can't really do that from the inside of the visit funcs.

This new Walk code is *much*, much cleaner and clearer than the previous generation it was ported from because all the issues about name normalization are now nicely and completely answered by standard use of RelPath.  While that meant the port was more work than copypasta, oh, does it feel nice.

Signed-off-by: Eric Myhre <hash@exultant.us>