Introduce tests mixin package, with file fixture typedefs and utility function for setting them up.
Went a little crazy enumerating almost every fixture variation I could think of, individually.  One of the todo's still visible in the current repeatr master branch tests is a lament about not covering individual variations to ensure their impacts on hashes, so, we're gonna focus on that detail right away this time.

Signed-off-by: Eric Myhre <hash@exultant.us>