Add spec tests for variations in filesets causing varations in hashes!
Add some additional fixtures for permissions variations, because ... sigh! ... not all permissions variations are the same!  Part of the fuckery we need to cover is the interplay of umask settings in the host causing changes in how our chmod calls are interpreted.  (File this as one of many reasons why it's important to have these tests run against a real filesystem with real syscalls, even though we went through plenty of effort to make that fs.FS interface theoretically mock'able.)

The osfs package sprouts a new syscall in its init function!  To fix... yup, you guessed it, that umask issue.  I don't usually condone the use of package init functions, but: A) the effect we're working with is literally program-global no matter what we might wish; thanks linux; and B) that package has *literally* one job; if you're importing it, you want this effect.

Signed-off-by: Eric Myhre <hash@exultant.us>