Fix reporting of non-file sizes.
These Odd semantics created fussery downstream in the tar transmat because it would cause us to say "hej tarWriter, we have 4096 bytes of content coming up FOR THIS DIRECTORY" and then not giving any such content because of course that doesn't exist.

I can't see any reason to confuse those two fields, so, just drop it.  If we do want to retain that "size of dir nodes" sort of chaff, I'd advocate for giving it a different struct field to avoid stupid semantic mixups like this one.

Signed-off-by: Eric Myhre <hash@exultant.us>