Uffdah.  The zero value of time is very problematic.
If I'm understanding it correctly, the zero value for time gets turned into an actually effect of *epoch* time on the filesystem, and then read back as that -- which is a very non-zero value.  I don't know if it's possible to avoid this lossyness, other than to studiously avoid ever using the zero value of time.

Signed-off-by: Eric Myhre <hash@exultant.us>