api/rio: Re-draft monitoring (chan!) and messages.
Channels are good for you.

Some review notes left in and decisions deferred for the further future.

Probably the biggest question in play in the near future is what to do with the error types.  We designed the `errcat` package *specifically* to be easy to round-trip serialize and be very schematically trivial... but that still doesn't mean we want to have it be an import of our API package (something we want as utterly minimal as possible, aside from the sheer table stakes of the serialization library itself).

Signed-off-by: Eric Myhre <hash@exultant.us>