api: Propose new Filters struct; docs.
This structure is simplified, because that other one is a MESS to serialize, and I don't think anyone wants it, and it was *still*, *also* a mess to handle in your program.

We also added a field to specify whether or not the sticky bits should be preserved on unpack, because some lark of research today showed that indeed, tar commands default to not footgunning you with those powerful properties.

The documentation now describes another enum value for some modes of operation which is designed to make operation proceed at the lowest privilege level possible ("mine").  This will be a new feature in this round of rio (and much appreciated, I hope -- it was requested often enough).

Documentation now covers all sorts of things more explicitly.

The down side of all this is that applications (well, by applications I guess I mean "rio" and rio alone, so really it's not so bad) now have to do a fair bit more parsing to make sense of the filters.  But this is no different than the situation with warehouse addresses -- so we should be fine.

Signed-off-by: Eric Myhre <hash@exultant.us>