cmd/rio: convert several flags to args.
They're essential to the meaning of the command and not optional nor plural so they gain nothing by being flags.

The placement of path *first* is contendable; the opinion I'm operating on for the moment is that it's a more useful indicator of the command's intent from the POV of someone watching `ps` than the wareID is (and also, the wareID itself is less than the whole story for that user, since any filter flags may alter the actual result of unpacking).

Signed-off-by: Eric Myhre <hash@exultant.us>