assembler: early draft.
This design attempting to proxy the UnpackFunc defns is... interesting, but I think flawed.  It's never sensible to stack (and that alone should be an overridingly powerful reason to stop, right there), and chock full of semantic drifts that our type system is simply too sloppy to make sufficiently apparent (e.g. the error returns).

Also considered was the possibility of making the assembler vary in cleverness for different placer funcs -- but this we will also probably abandon, because the one time we would care is for copy-only placement to operate without needing CAS cache areas and a potential double-copy, and that's... not something that's a huge priority to optimize for, especially seeing that in practice, even if you were to as a user story unpack all your development assets on your host disk... you'll still have a CI service 20 seconds later binding them all into a container to do your unit tests.  So, you indeed might as well have the CAS cache step.  It still serves its purpose in 99% of known user stories.

... If this commit message makes no sense to a reader, it's because it's referring to a bunch of code and concepts (e.g. "placer") from Repeatr which have not yet been ported over, but will be shortly.

So, long story short, this draft will immediately be dead-ended, and we're going to resort back to the fairly simplistic single assembler implementation that has served Just Fine in Repeatr for all this time, and revisit that again in the future if/when proven necessary (not before).

Signed-off-by: Eric Myhre <hash@exultant.us>