assembler: a port of the repeatr AssemblerPart struct.
Comments whitewashed a tad.  Some review todos emplaced about how we want this to function in the future.

It remains both tempting to try to get a mega object of code that does all unpacking and etc in one fell swoop from a giant dictionary of intentions, and yet fairly horrific to make an API that with that sprawling of a face.  How that tug 'o war shall be resolved remains to be seen.

Signed-off-by: Eric Myhre <hash@exultant.us>