Completely give up on the ErrFS type.
It got in the way just enough to force its expungement: it didn't *really* suitably replace the os.IsNotExist predicate for errors.

Why?  Because we both tried to get fancy with providing ~~meaningful information~~ (wow) like which path the error is about in our ErrNotExist type, but then wussed out on actually coercing things to it...

Why?  Because sometimes *the linux kernel doesn't give you a clue* which of the relevant paths is the source of the error.  So we *can't* making that meaningful info appear.

I don't know what to do about this except shake my fists at the stars and the linux kernel.

(Why?  One more layer of "why".  Probably because "returning" a path string from a syscall is kind of a no-go because there's no bound on how much memory that would require.  But ffs, why not a ENOEXISTS_ARG2 or something?)

So.  Fuckkit.  Now everything that we had which returns ErrNotExist returns os.NotExist.  Sane?  Not even remotely.  What else to do?  Idk.

Signed-off-by: Eric Myhre <hash@exultant.us>