stitch,cache: design work.
Making caches work *well* and compose *gracefully* (and then with an API we're actually *happy to support*) requires a surprisingly large amount of thought.

This was a similarly hard problem in the first generation of Repeatr's transmats, and we've been making subtly different choices this time around in Rio.  I kind of can't believe I just wrote a nearly 200-line document of rationale for the upcoming code (which honestly might be shorter than the document...), but it's tricky and I'm hopeful that this design and all this forethought will shape up even better this time around.

Signed-off-by: Eric Myhre <hash@exultant.us>