fsOp: add mkdirall convenience, and tests.
This will be handy for some of the cache system code.  (It's interesting to note that every other time we've had a recursive mkdir needed so far, it's had different/additional constraints, meaning this function will not in fact replace any of them.)

Tests so far only cover happy path; more to come.

Signed-off-by: Eric Myhre <hash@exultant.us>