cache: handle effect of filters on cacheability.
Which is to say, they break it.  For now.

(We can replace this with cleverer logic later; for now, this is much, much, much easier.  Recording tuples of {WareID,filters}->{resultWareID} in a way that we can upload to the same concurrency standard of the rest of the caching implementation -- that it must be safe with no collaboration mutexes other than standard posix filesystem atomicity -- is probably possible, but also certainly requires care.)

Signed-off-by: Eric Myhre <hash@exultant.us>