stitch/placer: outline how mount placer selected.
This is going to be implemented as a fairly alarming amount of reach-out to env vars, at least at first.  This branch for getting "assemblery" to work has already diverged a kilometer from master and I'm getting increasingly desperate to get it to MVP, even if hacky in its configurability, so we can merge and start moving on in bitesized pieces again.

Signed-off-by: Eric Myhre <hash@exultant.us>