stitch/placer: port bind placer.
Added RepairMtime helper to fsOp conveniences.  (This one is built with clever use of defers; its predecessor took a func arg.  This one is nicer -- because your stack isn't getting deeper, returning errors from the not-deeper logic doesn't get any more complicated.)

Some of the mkDest logic is heavy on the 'os' package rather than our 'fs.FS' abstraction layer.  This isn't a huge issue since when we're talking about mount placers, we're *very* much past the point of no return on cute abstractions, and entirely banking on the kernel and us having the same understanding of syscalls and their semiotics... but it's also mostly written as such out of laziness on my part in porting it, and this may be revised in the future.

Signed-off-by: Eric Myhre <hash@exultant.us>