stitch/placer: add cleanup funcs.
I was thinking about skipping this and attempting to replace it with a generic cleaner-upper func which would try to both recursive remove and detect and handle mounts as necessary along the way... but that's A) complicated, and B) seems like it might be slower from all the extra checking you would do.  We'll still maybe want such a function someday, simply to have as a helper for cleaning up on unhappy crashes and such, but that can come later; it's much easier to simply keep this info and DTRT when on the happy path.

Signed-off-by: Eric Myhre <hash@exultant.us>