stitch: possibly a working assembler.
Cleanup and error handling during it is... always tricky.  And I think that sentence deserves an award for understatement of the century.

Note that we dropped the ability to use a copy placer for an unpackTree here.  But not by much: pretty much the only thing in the way of making a mount-free system is A) rejecting "mount" type inputs, and B) replacing the 'placer.GetMountPlacer()' line with a CopyPlacer.  Though an actual *user-friendly* unpack-tree function might require rather more: for example, typically one often doesn't actually want to unpack the "/" item, even though it is typically necessary when specifying a container (and cleanup removal behavior may similarly need some guardrails).

Signed-off-by: Eric Myhre <hash@exultant.us>