stitch: better janitor failure handling and report
The hope is that this error message will actually provide enough information that a sysadmin can finish cleaning up carefully by hand without any serious questions.

The amount of trying is now also dramatically higher.  We step carefully around recursive removes (namely, in the copy placer) because of the possibility of incorrectly traversing a mount that failed to be unmounted.  But there's no reason *not* to try other unmounts, so we do.

This is probably the most verbose error report string in the whole project by miles, but this is also an error we expect to see rarely indeed, so it should be fine.

Signed-off-by: Eric Myhre <hash@exultant.us>