stitch: func to flip formula into []unpackSpec.
This is roughly the last bit of glue required before Repeatr can just load the formula, FormulaToUnpackTree, Assembler.Run(), and then bam, it's ready to invoke the container.

The prospective `rio unpack-tree` feature will probably internally use this same function, despite not bothering with many of the other fields in a formula -- it's useful to be able to say `rio unpack-tree formula-you-already-had.frm` and get the same tree (modulo some other sensible defaults, like probably skipping the "/" entry, etc).

Signed-off-by: Eric Myhre <hash@exultant.us>