stitch: return result map from PackMulti.
There's more info that's important to return from multiple packs than there is from a tree of unpacks: in unpacking a tree, we can basically forget about the wareIDs again because we typically only care about the whole set of filesets now manifested.  Packing is the opposite: we're done with the filesets, and its the WareID we care about after this.

Signed-off-by: Eric Myhre <hash@exultant.us>