transmat/tar: handle nonexistent targets; tests.
Also, improve the grace level of getting a string that's not an absolute path.  (We broadly plan to Not Have This Question because both the Rio CLI should absolutize asap, and Repeatr things get hyperexplicit early on too, but we might as well have defense/sanity-in-depth (especially since our declared types here really aren't providing any armor).

Signed-off-by: Eric Myhre <hash@exultant.us>