tar: emit binaries in stable order.
Mind you: the WareID hash we compute and return *is not affected* by this order, and by design never has been.

The binary blob of tar output we emit however, is affected: it is now stable and reproducible itself.  (Previously, it may have been defined in part by the filesystem iteration order, which is undefined on some filesystems.)  This should increase our niceness when our artifacts are consumed by other tools which do not see into the blob or use our concept of wareID hash (e.g. let's say someone wanted to feed the tars to diffoscope for some reason).

(Yes, this is a change in the fs package.  But the tar output is the only thing affected, so I'm labelling the commit "tar".)

Signed-off-by: Eric Myhre <hash@exultant.us>