fs: introduce ResolveLink.
There are a whoooole bunch of places where we want to be able to easily do inspections of file system state... as it will be seen from within an actual chroo, which this shifts the semantics of symlinks quite a lot.  We're going to centralize that logic in one place (and test the bejezus out of it).  Thus: the ResolveLink function.

Signed-off-by: Eric Myhre <hash@exultant.us>