cli: mega reorg.  Functionally mostly identical.
All subcommands are set up in blocks now, in which their arg value slots, parser defining calls, and action funcs are all visually cohabitated.  (This requires some slightly odd syntactical conventions, but it's worth it.)  The blocks also do scope limitation, and that *has* in fact fixed at least one small but embarassing hidden bug typo already.

I dropped the 'mirror' subcommand, since it wasn't wired to anything yet (although I'm sure it will be Soon now), and a bunch of other flags on the base command for the same reason.

A couple demuxing functions are now pulled out into their own files, though admittedly they don't exactly contain a huge range of options yet.  (This take strings rather than 'api.PackType' as arguments because it's whether they return errors or not that effectively defines whether those strings are fair game to cast to 'PackType'.  inb4 "wouldn't it be nice of go had enums"; yes yes it would.)

The error handling path for args parser errors jumped around a little.  It's now a bit more likely to emit '--format'-respecting errors even for args parsing errors -- though of course a mangled "--format" flag itself still won't leave you with a great result.

One major todo: jeez, the exit code mapping.  We're going to go put that in the upstream go-timeless-api repo, since we did so for repeatr, and we're still hewing to the idea that those are an API spec thing.

Signed-off-by: Eric Myhre <hash@exultant.us>