cli,rioeexecclient: fix all exit code handling.
We were previously falsely reporting a lot of zeros.  That's now fixed: we now exit with numbers from our table of documented exit codes.

And in the largest part of this diff... the client is adapted to A) not freak out for starters and B) now also actually confirm everything lines up with the serialized structural error messages.

(We've spawned another giant piece of copypasta compensating yet again for the fact the go stdlib is kind of a goober about admitting to exit codes.)

Signed-off-by: Eric Myhre <hash@exultant.us>