stitch: add rio.Monitor to {P,Unp}ackSpec.
The default is the zero monitor.  Users of the feature can reach in and set a monitor per spec if they desire it.  (There's no clear way to take a single monitor handle and mux it all transparently because progress messages don't have a demux-able field; we just didn't design it like that.)

Signed-off-by: Eric Myhre <hash@exultant.us>