fs,tar: SplitParent helper, and use case.
Fixes issue where tar unpack would not create an implicit "." dir.  Which, oddly, was *sometimes* fine: depending of course on whether the dir existed already -- in our unit tests, it usually did; but in real use with an empty cache, it might not.  So, now that's fixed and works sanely when using a cache: the tar unpack will always create that dir correctly.

Signed-off-by: Eric Myhre <hash@exultant.us>