transmat/tar: extract warehouse reader dial.
Partially.

Mirror will use this same function (when we add that -- which is soon, and why I'm here); scan, interestingly, not quite!

And we also still have the parsers in the warehouse packages themselves to deal with.

So there might be a new higher level abstraction from the warehouses and their parsers coming up soon.

Signed-off-by: Eric Myhre <hash@exultant.us>