transmat/tar: remember dirs handled during unpack.
Making nilfs memorize dir creation was a mistake.  This whole situation with tars allowing implicit parents is a problem that only tars have, and the logic to handle it should be confined to the tar package.  Moreover, it was a horrific code smell for the tar unpack to have been externalizing its logic for this all the way to filesystem state in the first place... and that has now been corrected: the stat calls here are gone.

Last commit didn't pass all tests.  I leave it as a monument to foolishness.

Signed-off-by: Eric Myhre <hash@exultant.us>