transmat/tar: fix mirror byte buffer copying.
Dire bug; was not even remotely correct in the face of reads shorter than the requested length, getting redundant and excess data on such short reads.

Our existing test coverage did not detect this because such short reads don't occur on a local filesystem.  Testing this is hard without getting a network in play (or, a lot of mocking -- either way, it's unpleasant).

Signed-off-by: Eric Myhre <hash@exultant.us>