transmat/tar: verify wareID AND return the filtered/actual unpack wareID.
Previously, a specifying filters that caused changes to the wareID would cause unpack/mirror to error, even when they're doing the right thing, because of over-zealous hash checks and failure to keep *both* sets of metadata.

Previously... this would be a problem for unprivileged/"mine" unpacks, as you might imagine.  Now, those unprivileged/"mine" unpacks work 100% the way you want them to.  Wahoo!

There's some redundancy here when the filter is *not* causing changes which would be easily avoided... todo:next

Signed-off-by: Eric Myhre <hash@exultant.us>