cmd: unpack: args order is revered to wareID, path
The path-then-wareID ordering was at some point chosen on the idea that it would make 'ps' output list the more "relevant" info first, but that turned out to be a pretty dumb idea in practice.  It's arguable what's "relevant" in the first place, undermined COMPLETELY by the fact the exec client will put both positional args at the end *anyway*, and on top of that the inconsistency with other commands is just surprising.

So.  Back to the order that was in fact original: wareID-then-path.

This might break a few shell scripts, but not too many should be out there at this stage.  And anyone consuming the exec client API will just instantly get the fix.

Signed-off-by: Eric Myhre <hash@exultant.us>