fs: fix stringing of relpaths starting with dot.
Previously, the stringer still produced values that are valid relative paths... but not the canonical forms we expected this method to produce.  This turns out to have caused issues downstream in the walk functions, and other nasty ramnifications.

So, now, parsing relpath and emitting it again for values like ".wonk" will emit "./.wonk".  As it should.

Signed-off-by: Eric Myhre <hash@exultant.us>