fix go-git tests.
The SanitizeRemote tests are too literal and fragile; I'm dropping them.
I assume upstream library has enough tests for those; if we're going to
carry any of them redundantly, I'd like it to be a very short whitelist
that covers only the features that we know we're very explicitly relying
on ourselves.

The only other fix is a tiny change to a string assert in error messages
(*also* arguably over-literal... though since the state of the art in
golang error handling is generally quite stringy, well... we try our best.)

Green!

Signed-off-by: Eric Myhre <hash@exultant.us>