Shared API update.  And, new filter specification.
This is kind of massive.  We're getting over to the 'redux' branch in
the go-timeless-api repo, which is where a bunch of things have been
redeveloped to be more suitable to the "stellar" project for pipelines.
Some changes are purely typographical.  Some are more serious.

'WarehouseAddr' was renamed to 'WarehouseLocation', which accounts for
an unfortunately large and noisey percentage of the diff.

Filters have been drastically reworked.  They are now specified as
structured strings on the CLI, like this:

  uid=1000,gid=1000,mtime=@1262304000,sticky=keep,setid=keep,dev=keep

or another example, but for unpack:

  uid=follow,gid=follow,mtime=follow,sticky=follow,setid=follow,dev=follow

Setuid/setgid bits will now be *rejected* by default rather than
quietly stripped; and sticky bits will be *kept* be default, rather
than stripped, since despite being in the same octets, sticky bits are
communicating properties that are quite safe.

Device nodes will also now be *rejected* by default.

All of these new filter behaviors are of course configurable.
The intention of these new defaults is to have normal packing be more
likely to create things which are easy to unpack without requiring
privileges on the unpacker side; so, while you *can* pack things with
setuid/setgid bits on, of course, you should have to opt-in as a form
of acknowledgement that what you're doing might be worth rethinking.

Note that the defaults when Repeatr uses Rio may be different
(Repeatr will generally keep more things, as e.g. setuid/setgid bits
used within the context of a container are rather different).

There's a *new error* returnable by Rio!  ErrFilterRejection -- to deal
with those nice new filter modes we're adding support for.

SkipChown behavior internally moved around.  It's now engaged
automatically when the uids and gids match.  This makes it apply more
easily, and fixes some internal oddly close coupling of distant config
to syscalls.  Generally things just "do the right thing" now.

The 'stitch' package is reduced in scope -- several "helpful" (cough)
methods that were essentially for working with repeatr formulas have
been removed.  They will reappear in repeatr.
(I think the original theory was that we might add some rio commands
which would consume the input section of formulas; however, this has
never materialized, and at this point I think we'd want to review
whether or not we still think that sort of responsiblity might belong
in the rio CLI at all.)

Since the go-timeless-api packages have now conceded to have more
helpful types (e.g. a Monitor object with methods as opposed to bare
channel), some logging code has become simplified.
(In the future this should probably go a step further and the relevant
methods should start returning log event objects rather than taking a
monitor parameter, but that was one step too much for bundling with
this already-sizable refactor.)

... Whew!  Just a few things.

Ah, and the exec-client package is also removed.  That will reappear,
and be updated for the new CLI, in the go-timeless-api repo presently.
Usage has increasingly shown that there are many consumers of the API
which would like to use this kind of client, but don't have any reason
to import the rest of the Rio git repo.  So, let's make it so.

Signed-off-by: Eric Myhre <hash@exultant.us>