Fix typo; and improve placefile messages for rejected filetypes.
Sockets already explain themselves well enough... though we *could*
implement creation of dead/unlistened socket files clearly enough.

Hardlinks... are more complicated.  I should make a longform writeup
about hardlinks for the docs.  Long story short, they're more or less
nondeterministic.  Hardlinks having any kind of *order* is a *tar*
invention; hardlinks cannot be coherently defined without first
defining the scope of the filesystem; and so forth.  It's not
*technically* difficult to implement hardlink support of some kind;
it's just that it's hard to say what correct and wise behavior *is*.

As a heuristic for correctness: I would say a formula which takes
some tarball with hardlinks as a single input at '/', and then
emits output for '/foo' and '/bar', both being half the filesystem
and having hardlinks stretching between them... such a formula should
have deterministic results, and should do so regardless of the order
of the hardlink pointers in the input tarball.

And if anybody can figure out an uncontrovertial spec for how to do
that, then PRs are most assuredly welcome.

Signed-off-by: Eric Myhre <hash@exultant.us>