Many small fixes and a few test fixture updates.
The helpful go1.10 auto-caching mechanism hid these from me earlier.

I'm a little salty about that.  I *know* these tests have impure
side effects; and I *know* they're dependent on state outside the
pure code version.  This is a system for *making syscalls*, for
god's sake.  Unless someone wants to bundle a purely functional mock
kernel within the view of the go test cacher, this just isn't
happening.  The attempt is bold but holy crap there needs to be a way
to opt out of this feature.  I don't want to bundle a "testme" script
that sets "-count=1" hack flags all the time!

Anyway.

- Update some end-to-end tests using the old filter flag syntax.

- Update some end-to-end test fixtures for serial messages to account
for the fact omitempty tags work correctly now (because of both refmt
features we've finally updated, and a typofix in the schema in the
go-timeless-api repo).

- Fix using the shorthand syntax in a filter initialization; that only
works if you also apply defaults afterwards, and in the test, that was
more verbose than just using a well-known one anyway.

And here's the most interesting one:

- Change all Event_* messaging to use non-pointer concrete types.
This is kind of an amusing fix, because we'd already written all the
*type* switches as if pointers would never be used, but then goofily
didn't fix the sending side to be non-pointer.  And unfortunately,
this is the sort of thing where trying to trick Go's interfaces into
acting like a proper sum type *doesn't work very well*: the pointer
types satisfy the interface just as well as the concretes, of course!
Using something like https://github.com/BurntSushi/go-sumtype would've
caught this (again; I feel like I leave a commit message of this form
behind me kind of a lot in the last year or so), but I keep saying I'm
going to set that up, and I keep not doing it; the friction of wiring
the same thing up for $N$ projects is just tool high.  I should
probably file this commit as an experience report to the Go team...

Signed-off-by: Eric Myhre <hash@exultant.us>