tar: log and skip 'g' entries.
As described in the comments in the diff: "in practice" this is always
the right thing to do.

Perhaps we'll refine that opinion later, but so far, this advisory info
is the only case I've ever seen; and evidently, 'g' headers aren't
often used by tar implementations for much else, since Go's tar impl
has never supported them, and I've never heard a lot of complaints
about that.

Frankly they're a silly concept in general and if we ever see content
out there which depends on the 'g' system to actually *do things* to
the tar content... I'm not sure we should support that.  It's just too
silly and too likely to be implemented buggily.

Signed-off-by: Eric Myhre <hash@exultant.us>