Fix default filters for scan to not alter hash.
Having the scan command do anything but "follow" or "reject" modes is
silly and results in scan producing hashes that... really aren't good
for anything, because you can't use those hashes for verifying the
content at the same source you just scanned without the filters again,
and that's just not how all this works.

This is back to being closer to how scan used to work; the different
defaults snuck in along with the other filter system changes in the API
refactor commits.

Signed-off-by: Eric Myhre <hash@exultant.us>