transmat/git: already-up-to-date is not an error!
I'm surprised we've never hit this before, but in practice I guess it's
been kind of rare: to encounter this path, you'd have to ask for a hash
you haven't asked for before, while also the refs you fetch are all
identical.  But of course that's possible!  And absolutely fine...

Signed-off-by: Eric Myhre <hash@exultant.us>